requirements.txt
websockets
mysql-connector-python
python-dotenv